Add initial support for TLS authentication with an HSM
This patch adds a few new arguments for registering a device using a
hardware security module (HSM) to back the TLS connection:

- --uuid: this allows providing a per-device UUID instead of always
  generating one randomly. That can be useful in this context if the UUID
  is associated with the HSM itself.

- --hsm-module: a shared object file implementing the PKCS#11 API
  for an HSM

- --hsm-so-pin: the Security Officer PIN to set on the HSM.

- --hsm-pin: the PIN to use on the HSM, and save in the aktualizr
    configuration file.

When --hsm-module is provided, --hsm-so-pin and --hsm-pin must also be
given. The two will be used to initialize a token and PIN-based login
on the HSM, then use that token to generate a per-device TLS key, then
form the TLS CSR using that key. The final certificate obtained
from the server is also stored in the HSM.

A corresponding aktualizr configuration file is also generated to use
these credentials when authenticating with the OTA server. This is
modified using the new "overrides" mechanism recently merged into the
Foundries.io OTA API.

Uptane key support is not yet provied as the pkcs11-tool utility used
to interact with the HSM does not support the ED25519 key type we
would like to use.

Signed-off-by: Marti Bolivar <marti@foundries.io>