Make writing of config files more safe
The way we write out config files isn't completely safe. There are a
couple of problems:

1) sota.toml can get written before other files like pkey.pem
If the device reboots, it might come up and *think* it's registered but
not actually have everything it needs.

2) We don't write files safely to disk
This could result in something like a zero-byte sota.toml.

This change introduces a safe way to write files. The same technique is
used in Fioconfig where you:

 * write to a temp file
 * fsync this to disk
 * rename the file to the actual file

This sequence is power fail safe.

Additionally, we only write sota.toml to disk after we've written
everything else.

Signed-off-by: Andy Doan <andy@foundries.io>