options: debug info: fix source of tag and factory definition
This commit fixes the debug information reporting the source of the tag
and factory definition.

There is one caveat: if a information provided by the command line
matches the information provided in a previous source (either the
environment or the configuration file), the former will be reported as
the original source of information.

This can be misleading but from the functional point of view it will be
irrelevant.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>