options: tee: manufacturing protection
This option allows lmp-device-register to interrogate the CAAM hardware
for an EC public key (256 bits usually)

The key will be stored in /var/sota in PEM format (for reference) and
passed to the gateway in the CSR using a custom extension [1]

Afer registration aktualizer-lite should periodically request the CAAM
hardware to sign random strings and send these signatures to the gateway
along with the message.

See [2] for optee-client reference.

Upon receiving these digests, the gateway shall verify them using the
board associated public key

[1] "1.3.6.1.4.1.294.1.00"
[2] https://github.com/OP-TEE/optee_client/pull/352

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>