main: add --force option to allow cleanup of previous execution data
The current implementation allows multiple key pairs to be stored in the
HSM using the same "tls" label, as long as multiple "client"
certificates. That leads to SSL errors inside aklite. Even a previous
incomplete lmp-device-register execution might leave keys behind.

This commit detects such conflicting objects, either aborting the
execution or removing them, if the new --force is set to 1. The new
option also enables the removal of existing sql.db and client.pem files,
allowing a re-registration to be performed using a single command.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>