main: cleanup created PKCS11 keys on registration errors
Even after succesfully contacting the registration server, the
process might still fail. For example, due to a device name conflict.
This leaves leave a key pair left in the HSM, which can lead to a
failure when retrying the registration.

This commit removes the created keys, allowing the command to be
re-executed without requiring the use of --force 1.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>